Corrected integer type used to store the number of messages.

+++ src/mbox.c
@@ -122,7 +122,7 @@ int mbox_time(const char *message, struct tm *time)
  *
  * @return Number of messages or -1 if there is insufficient memory
  */
-int mbox_read(char *mbox, char *messages[], unsigned long n)
+int mbox_read(char *mbox, char *messages[], int n)
 {
     char *p;     // pointer to next available mememory address in mbox
     size_t len;  // line length
@@ -131,7 +131,7 @@ int mbox_read(char *mbox, char *messages[], unsigned long n)
     char *line = NULL;
     size_t size = 0;
 
-    unsigned long i = 0;
+    int i = 0;
     while ((len = getline(&line, &size, stdin))) {
         if (len == -1) {
             free(line);
